+++ Uwuify.DiscordBot.WorkerService/Commands/UserCommands.cs
@@ -13,87 +13,85 @@ using System.Threading.Tasks;
 using Remora.Discord.API.Abstractions.Rest;
 using Remora.Rest.Core;
 using Uwuify.DiscordBot.WorkerService.Extensions;
-using Uwuify.DiscordBot.WorkerService.Models;
 using Uwuify.Humanizer;
 
-namespace Uwuify.DiscordBot.WorkerService.Commands
+namespace Uwuify.DiscordBot.WorkerService.Commands;
+
+public class UserCommands : LoggedCommandGroup<UserCommands>
 {
-    public class UserCommands : LoggedCommandGroup<UserCommands>
+    private readonly FeedbackService _feedbackService;
+
+    public UserCommands(ILogger<UserCommands> logger,
+        FeedbackService feedbackService,
+        ICommandContext ctx,
+        IDiscordRestGuildAPI guildApi,
+        IDiscordRestChannelAPI channelApi)
+        : base(ctx, logger, guildApi, channelApi)
     {
-        private readonly FeedbackService _feedbackService;
+        _feedbackService = feedbackService;
+    }
 
-        public UserCommands(ILogger<UserCommands> logger,
-            FeedbackService feedbackService,
-            ICommandContext ctx,
-            IDiscordRestGuildAPI guildApi,
-            IDiscordRestChannelAPI channelApi)
-            : base(ctx, logger, guildApi, channelApi)
-        {
-            _feedbackService = feedbackService;
-        }
+    [Command("uwuify")]
+    [CommandType(ApplicationCommandType.ChatInput)]
+    [Description("Convert your message into UwU")]
+    public async Task<IResult> UwuAsync([Description("text")] string text)
+    {
+        await LogCommandUsageAsync(typeof(UserCommands).GetMethod(nameof(UwuAsync)), text);
 
-        [Command("uwuify")]
-        [CommandType(ApplicationCommandType.ChatInput)]
-        [Description("Convert your message into UwU")]
-        public async Task<IResult> UwuAsync([Description("text")] string text)
-        {
-            await LogCommandUsageAsync(typeof(UserCommands).GetMethod(nameof(UwuAsync)), text);
+        var outputMsg = text.Uwuify();
 
-            var outputMsg = text.Uwuify();
+        _logger.LogDebug("{commandName} result: {msg}", nameof(UwuAsync), outputMsg);
 
-            _logger.LogDebug("{commandName} result: {msg}", nameof(UwuAsync), outputMsg);
-
-            var reply = await _feedbackService.SendContextualEmbedAsync(new Embed("Uwuify",
+        var reply = await _feedbackService.SendContextualEmbedAsync(new Embed("Uwuify",
                 Description: outputMsg,
                 Colour: new Optional<Color>(Color.PaleVioletRed)),
-                ct: CancellationToken);
+            ct: CancellationToken);
 
-            return reply.IsSuccess
-                ? Result.FromSuccess()
-                : Result.FromError(reply);
-        }
+        return reply.IsSuccess
+            ? Result.FromSuccess()
+            : Result.FromError(reply);
+    }
 
-        [Command("feedback")]
-        [CommandType(ApplicationCommandType.ChatInput)]
-        [Description("Leave feedback for the developer")]
-        public async Task<IResult> FeedbackAsync([Description("text")] string text)
-        {
-            await LogCommandUsageAsync(typeof(UserCommands).GetMethod(nameof(FeedbackAsync)), text);
+    [Command("feedback")]
+    [CommandType(ApplicationCommandType.ChatInput)]
+    [Description("Leave feedback for the developer")]
+    public async Task<IResult> FeedbackAsync([Description("text")] string text)
+    {
+        await LogCommandUsageAsync(typeof(UserCommands).GetMethod(nameof(FeedbackAsync)), text);
 
-            _logger.LogInformation("New feedback left by {userName}. Feedback: {feedbackText}", _ctx.User.ToFullUsername(), text.Trim());
+        _logger.LogInformation("New feedback left by {userName}. Feedback: {feedbackText}", _ctx.User.ToFullUsername(), text.Trim());
 
-            var reply = await _feedbackService.SendContextualEmbedAsync(new Embed("Feedback Submitted", 
+        var reply = await _feedbackService.SendContextualEmbedAsync(new Embed("Feedback Submitted", 
                 Description: "Thank you for your feedback! A developer will review your comments shortly.", 
                 Colour: new Optional<Color>(Color.PaleVioletRed)), 
-                ct: CancellationToken);
+            ct: CancellationToken);
 
-            return reply.IsSuccess
-                ? Result.FromSuccess()
-                : Result.FromError(reply);
-        }
+        return reply.IsSuccess
+            ? Result.FromSuccess()
+            : Result.FromError(reply);
+    }
 
-        [Command("Uwuify This Message")]
-        [CommandType(ApplicationCommandType.Message)]
-        public async Task<IResult> UwuThisMessageAsync()
-        {
+    [Command("Uwuify This Message")]
+    [CommandType(ApplicationCommandType.Message)]
+    public async Task<IResult> UwuThisMessageAsync()
+    {
 
-            var c = _ctx as InteractionContext;
-            var originalMessage = c!.Data.Resolved.Value.Messages.Value.Values.First().Content.Value;
+        var c = _ctx as InteractionContext;
+        var originalMessage = c!.Data.Resolved.Value.Messages.Value.Values.First().Content.Value;
 
-            await LogCommandUsageAsync(typeof(UserCommands).GetMethod(nameof(UwuThisMessageAsync)), originalMessage);
+        await LogCommandUsageAsync(typeof(UserCommands).GetMethod(nameof(UwuThisMessageAsync)), originalMessage);
 
-            var outputMsg = originalMessage.Uwuify();
+        var outputMsg = originalMessage.Uwuify();
 
-            _logger.LogDebug("{commandName} result: {msg}", nameof(UwuAsync), outputMsg);
+        _logger.LogDebug("{commandName} result: {msg}", nameof(UwuAsync), outputMsg);
 
-            var reply = await _feedbackService.SendContextualEmbedAsync(new Embed($"{c.Member.Value.Nickname.Value} just got UwU-ed!",
-                    Description: outputMsg,
-                    Colour: new Optional<Color>(Color.PaleVioletRed)),
-                ct: CancellationToken);
+        var reply = await _feedbackService.SendContextualEmbedAsync(new Embed($"{c.Member.Value.Nickname.Value} just got UwU-ed!",
+                Description: outputMsg,
+                Colour: new Optional<Color>(Color.PaleVioletRed)),
+            ct: CancellationToken);
 
-            return reply.IsSuccess
-                ? Result.FromSuccess()
-                : Result.FromError(reply);
-        }
+        return reply.IsSuccess
+            ? Result.FromSuccess()
+            : Result.FromError(reply);
     }
 }
