+++ Uwuify.DiscordBot.WorkerService/Models/ShortTermMemory.cs
@@ -1,10 +1,9 @@
 ﻿using System.Collections.Generic;
 using Remora.Rest.Core;
 
-namespace Uwuify.DiscordBot.WorkerService.Models
+namespace Uwuify.DiscordBot.WorkerService.Models;
+
+public static class ShortTermMemory
 {
-    public static class ShortTermMemory
-    {
-        public static HashSet<Snowflake> KnownGuilds { get; } = new();
-    }
+    public static HashSet<Snowflake> KnownGuilds { get; } = new();
 }
